+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2026-03-01] - PocketBase Loading Params Fix
+
+### Fixed
+
+- Fixed pocketbase store parameter update loop by properly tracking object equality and preventing redundant fetches. ([#86c8hant7](https://app.clickup.com/t/86c8hant7))
+- Fixed the `areObjectEntriesEqual` method to accurately compare objects with different keys or values. ([#86c8hant7](https://app.clickup.com/t/86c8hant7))
+
+---
+
 ## [2026-02-28] - View Transitions Service
 
 ### Added